feat: Enhance group management and session control with type standardization
This commit introduces new WebSocket-based group management capabilities and a unified integer type for IDs and timestamps across the application.

*   **Type Standardization**: Consolidates various ID and timestamp fields from `int32`/`int64` to the native `int` type across data models, protocol definitions, session management, and WebSocket handlers. This simplifies type handling and promotes consistency throughout the codebase.
*   **WebSocket Group Management**:
    *   Moves the `JoinGroup` functionality from a REST API endpoint to a WebSocket handler, improving real-time responsiveness for group membership changes.
    *   Implements a new `LeaveGroup` WebSocket handler, allowing users to gracefully exit chat groups.
*   **Logout Functionality**: Adds a `Logout` WebSocket handler to provide a robust mechanism for users to end their sessions and invalidate resume tokens.
*   **Internal Refinements**: Renames `ws.LeaveAllGroups` to `ws.ResetGroupScene` for clarity and updates the frontend to integrate with the new WebSocket-driven group join/leave and logout actions.

+++ internal/handler/api/avatar.go
@@ -14,7 +14,7 @@ import (
 )
 
 type avatarResp struct {
-	ID  int64  `json:"id"`
+	ID  int    `json:"id"`
 	URL string `json:"url"`
 }
 
@@ -41,7 +41,7 @@ func SetAvatar(c *gin.Context) {
 		return
 	}
 
-	userID := c.MustGet("user_id").(int64)
+	userID := c.MustGet("user_id").(int)
 
 	if err := user.UpdateAvatar(userID, req.AvatarId); err != nil {
 		log.Printf("[SetAvatar] UpdateAvatar user=%d avatar=%d error: %v", userID, req.AvatarId, err)
