+++ internal/model/avatar.go
@@ -1,6 +1,6 @@
 package model
 
 type Avatar struct {
-	ID       int64  `db:"id" json:"id"`
+	ID       int    `db:"id" json:"id"`
 	Filename string `db:"filename" json:"filename"`
 }
